requirements.txt
Django==3.1.7
djangorestframework==3.12.2
pylint==2.7.2